Add one-command runner and review fixes for the GPU run
- run_all.ps1 / run_all.sh: fresh clone -> setup (venv, CUDA torch, deps) ->
  TF-IDF baseline -> bi-encoder -> dense search -> union -> stage 1/2 ->
  compare -> stress -> predict -> France check; per-stage done markers,
  resume on re-run, UTF-8 console log, stops at the first error
- encoder: bf16 only on compute capability >= 8, tokenization prefetch
  thread, in-batch false-negative masking, float32 loss, config
  fingerprint for safe resume, atomic checkpoints, configurable
  intervals / weight decay / clip / gate k
- pairs: look-alike negatives (same name, then same first word) among
  encoder-fold S1s before falling back to random
- dense search refuses an untrained model and checks row universes
- union: derived absent-rank sentinel, int32 counts, config chunk size,
  resumable only when both outputs exist
- v5: early-stopping split per entity, training matrix freed after
  binning, predict uses the tau stage 2 was trained with, length checks
- env_check measures data sizes and exits without CUDA; probe takes any
  country; 54 tests; runner smoke-tested end to end on a CPU slice

+++ src/ber/features.py
@@ -64,12 +64,12 @@ def retrieval_features(cand: pd.DataFrame) -> pd.DataFrame:
     """Add record-side and S1-side retrieval features (numeric columns only)."""
     rec = cand.groupby(["src", "doc_row"], sort=False)["score"]
     cand["gap_rec"] = (rec.transform("max") - cand["score"]).astype(np.float32)
-    cand["n_cand_rec"] = rec.transform("size").astype(np.int16)
+    cand["n_cand_rec"] = rec.transform("size").astype(np.int32)
     s1src = cand.groupby(["s1_row", "src"], sort=False)["score"]
-    cand["rank_s1"] = (s1src.rank(ascending=False, method="first") - 1).astype(np.int16)
+    cand["rank_s1"] = (s1src.rank(ascending=False, method="first") - 1).astype(np.int32)
     cand["gap_s1"] = (s1src.transform("max") - cand["score"]).astype(np.float32)
-    cand["n_cand_s1_src"] = s1src.transform("size").astype(np.int16)
-    cand["n_cand_s1"] = cand.groupby("s1_row", sort=False)["score"].transform("size").astype(np.int16)
+    cand["n_cand_s1_src"] = s1src.transform("size").astype(np.int32)
+    cand["n_cand_s1"] = cand.groupby("s1_row", sort=False)["score"].transform("size").astype(np.int32)
     return cand
